Move cancel-visit-confirmation
